pyproject.toml
[project]
name = "search"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastparquet>=2024.11.0",
    "google-genai>=1.57.0",
    "ipykernel>=7.0.1",
    "langchain>=1.2.3",
    "opensearch-py>=3.0.0",
    "pandas>=2.3.3",
    "pyarrow>=22.0.0",
    "wikiextractor>=3.0.6",
]
